agentic_data_processor/requirements.txt
openai
langchain
langchain-openai
pandas
pyarrow
sqlalchemy
boto3
pyyaml
python-dotenv
tqdm
google-cloud-bigquery
streamlit
faiss-cpu